fix(op-deployer): EnsureDefaultCacheDir
