+++ op-deployer/pkg/deployer/utils_test.go
@@ -1,0 +1,12 @@
+package deployer
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/require"
+)
+
+func TestEnsureDefaultCacheDir(t *testing.T) {
+	cacheDir := EnsureDefaultCacheDir()
+	require.NotNil(t, cacheDir)
+}
